PodcastPlatform/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-redis==5.3.0
django-request-logging==0.7.5
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
django-rosetta==0.9.9
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
frozenlist==1.4.0
geoip2==4.7.0
gevent==23.9.1
greenlet==3.0.0
grequests==0.7.0
idna==3.4
kombu==5.3.2
maxminddb==2.4.0
multidict==6.0.4
packaging==23.2
pika==1.3.2
polib==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rabbitmq==0.2.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.17
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
